Added messaging projects and minor clean up

+++ CDC/SqlServer_CDC_Demo/ConsumerApplication/CDC.Loan/CDCRecords/PropertyCDCRecord.cs
@@ -1,8 +1,5 @@
 ﻿using CDCOutboxSender;
-using System;
-using System.Collections.Generic;
 using System.Data;
-using System.Text;
 
 namespace CDC.Loan
 {
@@ -18,14 +15,14 @@ namespace CDC.Loan
         {
             PropertyId = ExtractValueFromReaderIfPresent<int>(reader, "PropertyId");
             Address = ExtractValueFromReaderIfPresent<string>(reader, "Address");
-            Value = ExtractValueFromReaderIfPresent<int>(reader, "Value");
+            Value = ExtractValueFromReaderIfPresent<decimal>(reader, "Value");
 
         }
 
 
         public int PropertyId { get; private set; }
         public string Address { get; private set; }
-        public int Value { get; private set; }
+        public decimal Value { get; private set; }
 
     }
 }
